Build complete. Adjust logic  for FCC tests.
Although the site was fully functional it took a long time to resolve an
issue it was having with Free Code Camp's test suite. In the end by
making some changes to the execution of the alarm sound and by
changing the way the mm:ss rendering was achieved (I had to
look at the execution on the example site to see if anything they
did affect the way the test suite was executing) I managed to
get the code past the one test it was failing. It still doesn't seem
to be perfect as the setInterval handler was taking >50ms,
but I'm not sure why as it was only occasional.